Adds parsing of unbraced alphanumeric strings
An update was made to the way DOI references are returned in the
BibTeX format. Previously, all field values were returned in braces.
However, in November 2023, a change was made causing the month values
to be returned unbraced and unquoted. This allows them to operate as
macros when building a bibliography with LaTeX/BibTeX; however, it
was causing the btx <doi> command to fail. Normally, BibTeX field
values represented by a single plain number can also be encoded in
the unbraced and unquoted format. Therefore, a change was made to the
btx BibTeX parser to allow field values consisting of a single alpha-
numeric string (i.e., only 0-9, a-z, A-Z) to be recognized as valid
field values. This only affects the parser. All values are saved in
the braced format after parsing.